==> src/notebooks/03_model_types.ipynb <==
import warnings
from transformers import AutoConfig
warnings.filterwarnings("ignore")



MODEL_NAMES = [
    "j-hartmann/emotion-english-distilroberta-base",
    "bhadresh-savani/distilbert-base-uncased-emotion",
    "mrm8488/t5-base-finetuned-emotion",
    "SamLowe/roberta-base-go_emotions",
    "cardiffnlp/twitter-roberta-base-emotion"
]

for model in MODEL_NAMES:
    config = AutoConfig.from_pretrained(model)
    print(f"Architecture for {model}: {config.architectures}")
    print(f"Model type for {model}: {config.model_type}")
    print(f"Number of labels for {model}: {config.num_labels}")
    for label in config.id2label.values():
        print(f"- Label {label}")   
    print()
